chaos: refactor
To improve the maintainability of the model with a file for each class.
Added tests.
