+++ src/core/index.ts
@@ -1,4 +1,4 @@
 export * from './attribute.js';
-export * from './models.js';
+export * from './models/index.js';
 export * from './create-root-graph.js';
 export * from './to-dot.js';
