+++ src/core/models/DotObject.ts
@@ -1,0 +1,5 @@
+/**
+ * Base class for DOT objects.
+ * @group Models
+ */
+export abstract class DotObject {}
